Update #2 -- Stephen
Ok, for some reason Git isn't recognizing it when I make changes locally. So here's a reupload of the entire project, with the feather added and some rudimentary starting controls. It's currently hard as fuuuuuuck and I think we wanna rework the feather to give it a bit more room to expand... but it's kind of fun trying to balance both for now